implist: Add static_assert for minimum size and alignment
Instantiations must ensure their requested total size is double-word
aligned. They must also ensure there is enough space to fit at least one
block.

The first and last words in m_data are always reserved. The second and
second to last words are the header and footer of the initial free
block, respectively. Finally, the minimum aligned alloc size is a single
double-word. This results in a minimum requestable stalloc_t size of
3 * DSIZE.

Note that the minimum aligned size of a block is 2 * DSIZE, as this does
not include the reserved first and last words in m_data for the minimum
requestable stalloc_t size.