explist: Initial explicit list implementation
Features:

- Bidirectional bounding tags
- Bidirectional immediate coalescing
- Templated first fit or best fit via if constexpr
- Templated LIFO order or address order policy via if constexpr

The implicit list implementation requires iterating over both allocated
and non-allocated blocks to find an adequate fit on allocation. An
explcit list, on the other hand, stores a linked list of the free
blocks, meaning an allocation can iterate solely over the free blocks to
find an adequate fit. Since each free block is at minimum a double-word
size (16B on 64-bit architectures), we can store two pointers in the
first double-word of all free blocks; one pointing to the previous free
block node, one pointing to the next.

This implementation adds a new template parameter to specify the free
list ordering policy at compile time/instantiation. This defaults to
LIFO ordering (stalloc_ord_t::lifo_order).

LIFO ordering requires pushing the block to the head of the free list
when that block is freed. This means the last block freed may be the
first block popped off the free list on the subsequent allocation (this
is not guaranteed, however, since the free block may be too small).

Address ordering (stalloc_ord_t::addr_order) requires each block being
freed to be inserted into the free list in address order. This may lead
to less fragmentation, but also requires a variable free list walk on
each free.

Note that a best fit policy negates the benefits of address ordering.
When using best fit, it does not matter the ordering of the free list;
the 'best fit' will be chosen every time. The only time this isn't quite
the case is when multiple free blocks are the 'best fit', in which case
the current algorithm selects the first 'best fit' found, but this is
relatively negligible and difficult to reason about from a
performance/fragmentation perspective.

As such, using best fit in combination with address ordering leads to
needless free list walks, and is statically disallowed by default.